# KAFKA-14578: Move ConsumerPerformance to tools