… core/src/main/scala/kafka/tools/ConsumerPerformance.scala
… core/src/main/scala/kafka/tools/PerfConfig.scala
… core/src/test/scala/unit/kafka/tools/ConsumerPerformanceTest.scala