@fvaleri 2023-02-23
@mimaison thanks for the review.

I fixed the rebalance listener and re-ran all the test in the description.